Plein de trucs

+++ src/main/java/fr/maif/parsedontvalidatejava/v1/domain/model/Colis.java
@@ -8,30 +8,34 @@ import javax.validation.constraints.NotNull;
 import java.time.LocalDateTime;
 import java.util.Objects;
 
-@Data
-@Builder(toBuilder = true)
-public class Colis {
-    public String reference;
-    @NotNull
-    public TypeColis type;
-    @NotNull
-    public LocalDateTime dateDEnvoi;
-    public LocalDateTime dateReception;
-    public Integer latitude;
-    public Integer longitude;
-    @Email
-    @NotNull
-    public String email;
-    public Adresse adresse;
+ @Data
+ @Builder(toBuilder = true)
+ public class Colis {
+     public String reference;
+     @NotNull
+     public TypeColis type;
+     @NotNull
+     public LocalDateTime dateDEnvoi;
+     public LocalDateTime dateReception;
+     public Double latitude;
+     public Double longitude;
+     @Email
+     @NotNull
+     public String email;
+     public Adresse adresse;
 
-    public Colis(String reference, TypeColis type, LocalDateTime dateDEnvoi, LocalDateTime dateReception, Integer latitude, Integer longitude, String email, Adresse adresse) {
-        this.reference = reference;
-        this.type = type;
-        this.dateDEnvoi = Objects.requireNonNullElse(dateDEnvoi, LocalDateTime.now());
-        this.dateReception = dateReception;
-        this.latitude = latitude;
-        this.longitude = longitude;
-        this.email = email;
-        this.adresse = adresse;
-    }
-}
+     public Colis(String reference, TypeColis type, LocalDateTime dateDEnvoi, LocalDateTime dateReception, Double latitude, Double longitude, String email, Adresse adresse) {
+         this.reference = reference;
+         this.type = type;
+         this.dateDEnvoi = Objects.requireNonNullElse(dateDEnvoi, LocalDateTime.now());
+         this.dateReception = dateReception;
+         this.latitude = latitude;
+         this.longitude = longitude;
+         this.email = email;
+         this.adresse = adresse;
+     }
+ }
+
+
+
+
